changed my mind this shit works now

+++ RSDKv5/Math.cpp
@@ -32,9 +32,9 @@ void CalculateTrigAngles()
     randKey = rand();
 
     for (int i = 0; i < 0x400; ++i) {
-        sinVal1024[i]  = (int)(sinf((i / 512.0) * M_PI) * 1024.0);
-        cosVal1024[i]  = (int)(cosf((i / 512.0) * M_PI) * 1024.0);
-        tanVal1024[i]  = (int)(tanf((i / 512.0) * M_PI) * 1024.0);
+        sinVal1024[i]  = (int)(sin((i / 512.0) * M_PI) * 1024.0);
+        cosVal1024[i]  = (int)(cos((i / 512.0) * M_PI) * 1024.0);
+        tanVal1024[i]  = (int)(tan((i / 512.0) * M_PI) * 1024.0);
         aSinVal1024[i] = (int)((asin(i / 1023.0) * 512.0) / M_PI);
         aCosVal1024[i] = (int)((acos(i / 1023.0) * 512.0) / M_PI);
     }
@@ -43,15 +43,16 @@ void CalculateTrigAngles()
     cosVal1024[0x100] = 0;
     cosVal1024[0x200] = -0x400;
     cosVal1024[0x300] = 0;
+
     sinVal1024[0x000] = 0;
     sinVal1024[0x100] = 0x400;
     sinVal1024[0x200] = 0;
     sinVal1024[0x300] = -0x400;
 
     for (int i = 0; i < 0x200; ++i) {
-        sinVal512[i]   = (int)(sinf((i / 256.0) * M_PI) * 512.0);
-        cosVal512[i]   = (int)(cosf((i / 256.0) * M_PI) * 512.0);
-        tanVal512[i]   = (int)(tanf((i / 256.0) * M_PI) * 512.0);
+        sinVal512[i]   = (int)(sin((i / 256.0) * M_PI) * 512.0);
+        cosVal512[i]   = (int)(cos((i / 256.0) * M_PI) * 512.0);
+        tanVal512[i]   = (int)(tan((i / 256.0) * M_PI) * 512.0);
         aSinVal512[i]  = (int)((asin(i / 511.0) * 256.0) / M_PI);
         aCosVal512[i]  = (int)((acos(i / 511.0) * 256.0) / M_PI);
     }
@@ -76,7 +77,7 @@ void CalculateTrigAngles()
     for (int y = 0; y < 0x100; ++y) {
         byte *arcTan = (byte *)&atanVal256[y];
         for (int x = 0; x < 0x100; ++x) {
-            *arcTan = (byte)(atan2f(y, x) * 40.743664f);
+            *arcTan = (byte)(atan2(y, x) * 40.743664);
             arcTan += 0x100;
         }
     }
